Add InChI AuxInfo to structure functionality;
